requirements.txt
junos-eznc
fastmcp
PyYAML
lxml